+++ scripts/run-with-more-memory.sh
@@ -22,3 +22,6 @@ NODE_OPTIONS="--max-old-space-size=8192" node "$SCRIPT_PATH" "$@"
 
 
 
+
+
+