+++ supabase/migrations/20250203000004_fix_balance_updates_and_rls.sql
@@ -211,3 +211,6 @@ COMMENT ON FUNCTION trigger_reward_meeting_confirmed IS 'Trigger function that a
 
 
 
+
+
+