+++ supabase/migrations/20250203000006_backfill_rewards_for_existing_meetings.sql
@@ -154,3 +154,6 @@ COMMENT ON FUNCTION backfill_meeting_rewards IS 'One-time function to backfill L
 
 
 
+
+
+